updated to version 0.9.7

+++ src/builder/utils/empty_dir.rs
@@ -3,16 +3,16 @@ use std::fs;
 use std::path::Path;
 
 pub fn empty_dir(dir: &Path) -> Result<()> {
-  if let Ok(exists) = dir.try_exists() {
-    if exists {
-      for entry in dir.read_dir()? {
-        let entry = entry?;
-        let path = entry.path();
-        if path.is_dir() {
-          fs::remove_dir_all(path)?;
-        } else {
-          fs::remove_file(path)?;
-        }
+  if let Ok(exists) = dir.try_exists()
+    && exists
+  {
+    for entry in dir.read_dir()? {
+      let entry = entry?;
+      let path = entry.path();
+      if path.is_dir() {
+        fs::remove_dir_all(path)?;
+      } else {
+        fs::remove_file(path)?;
       }
     }
   }
